Added a footer to the page and an about page
The about page was for the sake of it, but it will be useful in future. It will need expanding towards the end of the project.

I've also put in links for svn and git help, with the idea of putting a list of the simple commands there later for quick reference