Removed ACL
It was a bitch and I don't like it any more. We had an argument. It won. So I've removed it from the project. I'm not bitter. All all. We just want to go our separate ways. It's not my fault she's a mother bitch that takes a genius to configure to work on anything but their example.