Complete (fingers crossed) the move to the new logging mechanism.
Changes are now logged to a central table where history can be drawn from.
This will allow us to show the changes that are made to Objects (i.e accountability)

The language files are now split into controllers as it would otherwise be one massive file.
This allows us to keep track of things we have internationalised.